+++ dropwizard-config-updater/src/main/java/ru/vyarus/yaml/updater/dropwizard/ConfigUpdaterBundle.java
@@ -3,6 +3,7 @@ package ru.vyarus.yaml.updater.dropwizard;
 import io.dropwizard.Configuration;
 import io.dropwizard.ConfiguredBundle;
 import io.dropwizard.setup.Bootstrap;
+import io.dropwizard.setup.Environment;
 import ru.vyarus.yaml.updater.dropwizard.cmd.UpdateConfigCommand;
 
 /**
@@ -14,5 +15,11 @@ public class ConfigUpdaterBundle implements ConfiguredBundle<Configuration> {
     @Override
     public void initialize(final Bootstrap<?> bootstrap) {
         bootstrap.addCommand(new UpdateConfigCommand());
+    }
+
+    // added for dropwizard 1 compatibility
+    @Override
+    public void run(final Configuration configuration, final Environment environment) throws Exception {
+        // empty
     }
 }