+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+* [dropwizard]
+  - Fix dropwizard 1.x compatibility
+  - Make dropwizard dependency "provided" (to avoid version clashes)
+
 ### 1.0.0 (2021-08-29)
 * Initial release